core: migrate episode ID from long to int
I initially thought that long in java is unsigned. Thus since ID of
an episode should always be positive it made sense to use "unsigned"
long.
Turns out long is just like int but bigger. Thus for simplicity reasons
and because int value range is always enough to enumerate all episodes
of a show - change episode ID type from long to int.

+++ core/src/main/java/org/serio/core/shows/Shows.java
@@ -90,8 +90,8 @@ public class Shows {
      * @param episodeId ID of the episode, the user is watching
      * @param watchProgress amount of the specified episode's duration the user has just watched
      */
-    public void watchShowEpisode(UUID showId, long episodeId, WatchProgress watchProgress) {
-        watchHistory.watchShowEpisode(showId.toString(), Long.toString(episodeId), watchProgress);
+    public void watchShowEpisode(UUID showId, int episodeId, WatchProgress watchProgress) {
+        watchHistory.watchShowEpisode(showId.toString(), Integer.toString(episodeId), watchProgress);
     }
 
     /**
